Use smallint for frequencies,

+++ mdh_corpus/models.py
@@ -2,6 +2,12 @@ from django.db import models
 
 # Create your models here.
 
+# TODO: optimise data types (e.g. PositiveSmallIntegerField for freq)
+# TODO: don't use postgresql autoincrement ID for m2m
+# https://docs.djangoproject.com/en/2.1/ref/models/fields/
+# #positivesmallintegerfield
+# https://www.postgresql.org/docs/current/static/datatype-numeric.html
+
 
 class Journal(models.Model):
     label = models.CharField(max_length=200, unique=True)
@@ -33,7 +39,7 @@ class Ngram1(models.Model):
 class Ngram1Article(models.Model):
     ngram1 = models.ForeignKey(Ngram1, on_delete=models.CASCADE)
     article = models.ForeignKey(Article, on_delete=models.CASCADE)
-    freq = models.IntegerField(default=0)
+    freq = models.PositiveSmallIntegerField(default=0)
 
 
 class Ngram2(models.Model):
@@ -43,7 +49,7 @@ class Ngram2(models.Model):
 class Ngram2Article(models.Model):
     ngram2 = models.ForeignKey(Ngram2, on_delete=models.CASCADE)
     article = models.ForeignKey(Article, on_delete=models.CASCADE)
-    freq = models.IntegerField(default=0)
+    freq = models.PositiveSmallIntegerField(default=0)
 
 
 class Ngram3(models.Model):
@@ -53,4 +59,4 @@ class Ngram3(models.Model):
 class Ngram3Article(models.Model):
     ngram3 = models.ForeignKey(Ngram3, on_delete=models.CASCADE)
     article = models.ForeignKey(Article, on_delete=models.CASCADE)
-    freq = models.IntegerField(default=0)
+    freq = models.PositiveSmallIntegerField(default=0)
